Move TableConstraint to Constraints conversion
Reduce datafusion-common dependency on sqlparser
